// packages/core/src/runtime/router/types.ts
export type TreeRoutingCtx<T = unknown> = Readonly<{
  treeId: string;
  flatNodes: readonly FlattenedNode<T>[];
  expanded: readonly string[];
  state: TreeLocalState;
  keyboardNavigation: boolean;
}>;